refactor: extract profile sidebar group
