+++ frontend/src/components/SidebarToggle/SidebarToggle.tsx
@@ -1,23 +1,8 @@
 import "./SidebarToggle.css"
 import { createPortal } from "react-dom"
 import { memo, ReactPortal, useState } from "react"
-import { toast } from "sonner"
-import {
-  LuAlignJustify,
-  LuHistory,
-  LuLogOut,
-  LuPodcast,
-  LuRadioTower,
-  LuRedo2,
-  LuUser,
-  LuUserPlus,
-} from "react-icons/lu"
-import { MdFollowTheSigns } from "react-icons/md"
+import { LuAlignJustify, LuPodcast, LuRadioTower } from "react-icons/lu"
 import { GoStarFill } from "react-icons/go"
-import {
-  signOut,
-  useSessionContext,
-} from "supertokens-auth-react/recipe/session/index"
 import useClickOutside from "../../hooks/useClickOutside.ts"
 import Button from "../ui/button/Button.tsx"
 import FavouriteStationToggle from "../../features/favourite/FavouriteStationToggle/FavouriteStationToggle.tsx"
@@ -26,28 +11,13 @@ import Sidebar, {
   SidebarMenu,
   SidebarMenuItem,
 } from "../Sidebar/Sidebar.tsx"
-import {
-  homePage,
-  podcastHomePage,
-  profileFollowingPage,
-  profileHistoryPage,
-  profilePage,
-  resetPasswordPage,
-  signInPage,
-  signUpPage,
-} from "../../paths.ts"
+import ProfileSidebarGroup from "../../features/sidebar/ProfileSidebarGroup.tsx"
+import { homePage, podcastHomePage } from "../../paths.ts"
 
 const homePageUrl = homePage()
 const podcastHomePageUrl = podcastHomePage()
-const profileSignInPageUrl = signInPage()
-const profileSignUpPageUrl = signUpPage()
-const profilePageUrl = profilePage()
-const profileFollowingPageUrl = profileFollowingPage()
-const profileListenHistoryPageUrl = profileHistoryPage()
-const profileResetPasswordPageUrl = resetPasswordPage()
 
 function SidebarToggle() {
-  const session = useSessionContext()
   const [open, setOpen] = useState<boolean>(false)
   const [portals, setPortals] = useState<ReactPortal[] | []>([])
   const clickOutsideRef = useClickOutside<HTMLDivElement>({
@@ -62,11 +32,6 @@ function SidebarToggle() {
       return
     }
     setOpen(false)
-  }
-  async function handleLogout() {
-    await signOut()
-    setOpen(false)
-    toast.info("SEE YOU SPACE COWBOY...")
   }
   function handleFavouriteStationClick() {
     setOpen(false)
@@ -114,63 +79,7 @@ function SidebarToggle() {
                 data-testid="sidebar-menu-item-radio-view-favourite-stations"
               />
             </SidebarGroup>
-            <SidebarGroup label="Profile">
-              {session.loading || !session.doesSessionExist ? (
-                <>
-                  <SidebarMenuItem
-                    title="Sign In"
-                    url={profileSignInPageUrl}
-                    Icon={LuUser}
-                    onClick={handleToggle}
-                    data-testid="sidebar-menu-item-profile-sign-in"
-                  />
-                  <SidebarMenuItem
-                    title="Sign Up"
-                    url={profileSignUpPageUrl}
-                    Icon={LuUserPlus}
-                    onClick={handleToggle}
-                    data-testid="sidebar-menu-item-profile-sign-up"
-                  />
-                </>
-              ) : (
-                <>
-                  <SidebarMenuItem
-                    title="Your Profile"
-                    url={profilePageUrl}
-                    Icon={LuUser}
-                    onClick={handleToggle}
-                    data-testid="sidebar-menu-item-profile"
-                  />
-                  <SidebarMenuItem
-                    title="Following"
-                    url={profileFollowingPageUrl}
-                    Icon={MdFollowTheSigns}
-                    onClick={handleToggle}
-                    data-testid="sidebar-menu-item-profile-following"
-                  />
-                  <SidebarMenuItem
-                    title="Listen History"
-                    url={profileListenHistoryPageUrl}
-                    Icon={LuHistory}
-                    onClick={handleToggle}
-                    data-testid="sidebar-menu-item-profile-listen-history"
-                  />
-                  <SidebarMenuItem
-                    title="Reset Password"
-                    url={profileResetPasswordPageUrl}
-                    Icon={LuRedo2}
-                    onClick={handleToggle}
-                    data-testid="sidebar-menu-item-profile-reset-password"
-                  />
-                  <SidebarMenuItem
-                    title="Logout"
-                    Icon={LuLogOut}
-                    onClick={handleLogout}
-                    data-testid="sidebar-menu-item-profile-logout"
-                  />
-                </>
-              )}
-            </SidebarGroup>
+            <ProfileSidebarGroup setOpen={setOpen} />
           </SidebarMenu>
         </Sidebar>
       </div>
